REF: share _simple_new
